perf(rvt): cut CUDA host->device transfer (per-window prefix + int32 coords)
- Kernel reconstructs each event's window in-kernel via binary search of a tiny per-window
  prefix-sum, instead of transferring membership-expanded index arrays (~2x event-data saved).
- Transfer x/y/p as int32 (their native h5 dtype) rather than int64 (halves that transfer).
- pipeline reads int32 coords for the cuda backend; drop evlib_gpu (non-UVM device pool) from
  the benchmark (UVM strictly dominates it).

987M-event sequence 28.2s -> 19.3s; 127M 7.1s -> 5.8s (RVT-GPU 18.7s / 8.4s). Bit-identical.

+++ src/ev_representations/mod.rs
@@ -86,9 +86,9 @@ pub mod python {
     pub fn stacked_histogram_dense_cuda_py<'py>(
         py: Python<'py>,
         t: PyReadonlyArray1<i64>,
-        x: PyReadonlyArray1<i64>,
-        y: PyReadonlyArray1<i64>,
-        p: PyReadonlyArray1<i64>,
+        x: PyReadonlyArray1<i32>,
+        y: PyReadonlyArray1<i32>,
+        p: PyReadonlyArray1<i32>,
         grid: PyReadonlyArray1<i64>,
         delta_t_us: i64,
         nbins: usize,
